move settings away from index.ts
